稳定 TB6612 任务模式并切到 ITEM2
将任务模式切到 TB6612 后端，关闭 Modbus 电机板专属握手流程；加入稳定的 OLED P1/P2 调试页；写入实测轮径、CPR 和直线 yaw 参数；修复编码器时间戳前跳导致的误报故障；补充安全回归测试和转向/ITEM1 调试记录。

+++ core/safety_supervisor.c
@@ -4,7 +4,8 @@ static bool CarSafety_IsStale(uint32_t now_ms,
                               uint32_t timestamp_ms,
                               uint32_t timeout_ms)
 {
-    return (uint32_t)(now_ms - timestamp_ms) > timeout_ms;
+    /* A live sampler can timestamp one tick after the caller snapshots now. */
+    return (int32_t)(now_ms - timestamp_ms) > (int32_t)timeout_ms;
 }
 
 uint32_t CarSafety_Evaluate(const CarConfig *config,
